update deploy base command

+++ Makefile
@@ -12,6 +12,16 @@ stop:
 	docker-compose -f docker-compose.yml stop
 
 deploy-base:
+ifeq ($(app), )
+	@echo "\033[0;32mPlease provide a repository name by writing: make deploy-base app=<value>\033[0m"
+else
+	@echo "\033[0;32mCreating a new WP repository: "$(name)"...\033[0m"
+	cd www && mkdir $(name) && cd $(name) && git clone https://github.com/$(GIT_USER)/wp-site-boilerplate.git && cd wp-site-boilerplate; rm -rf .git && mv .[^.]* .. && mv * ..
+	cd www/$(name) && rm -rf wp-site-boilerplate
+	cd www/$(name) && curl -i -H "Authorization: token "$(AUTH_TOKEN)"" -d '{ "name":"'$(name)'" }' https://api.github.com/user/repos
+	cd www/$(name) && git init; git remote add origin https://github.com/$(GIT_USER)/$(name).git
+	cd www/$(name) && git add . && git commit -a -m "Imported files from wp-site-boilerplate.git" && git push -u origin master
+	@echo "\033[0;32mSuccessfully created "$(name)"\033[0m"
 	rm -rf www/$(app)
 	mkdir www/$(app)
 	@echo "\033[0;32mCloning "$(app)" repository from GitHub ($(GIT_USER))...\033[0m"
@@ -22,6 +32,7 @@ deploy-base:
 	cd www/$(app) && make db-import-prod
 	@echo "\033[0;32mSuccessfully deployed "$(app)" to your local airbase environment.\033[0m"
 	@echo "\033[0;32mAccess URL: http://"$(app)".test\033[0m"
+endif
 
 stop-dnsmasq:
 	sudo brew services stop dnsmasq
